frontend: Extract NavigationTabs from Sidebar
